Add GitHub workflow that smoke-builds all extensions
This is known to pass, today, on x86_64 and fail on aarch64 due to
https://github.com/eclipse/hawkbit-extensions/issues/84

Since the goal is to test unreleased HawkBit + unreleased HawkBit extensions,
both are built in sequence to satisfy dependencies that are not available
through maven as releases.

HawkBit tests are skipped to speed up the build process.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>