s3: implement artifact range requests
The basic itent is to allow expressing an input stream for a fragment that
corresponds to a single range request.

This patch is not ready to land, because it requires lock-step update in
HawkBit proper. The long story short, is that the current abstraction of
getFileInputStream is just insufficient to efficiently implement range requests
for artifacts.

There are two goals for efficiency:

- We should not request or discard data we do not need
- We read all the data we requested and close the stream properly

The second problem is somewhat elusive, since S3 input stream has a close()
method that informs us of incorrect usage if not all content is drained, when
S3 is sending the content but are are not reading.

Corresponging HawkBit change is posted to HawkBit in a separate patch.

Fixes: https://github.com/eclipse/hawkbit-extensions/issues/91
Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>